modify
git-svn-id: https://svn.baidu.com/app/search/sep/trunk/anti-spam/gezi@10735 fa64baa9-71d1-4fed-97ae-c15534abce97

+++ include/Numeric/Vector/Vector.h
@@ -324,7 +324,7 @@ namespace gezi {
 
 			int nonZeroNum = GetNumNonZeros();
 			//@TODO 检查特征全是0的情况
-			if (nonZeroNum < (uint64)(length * maxSparsity))
+			if ((uint64)nonZeroNum < (uint64)(length * maxSparsity))
 			{
 				ToSparse();
 			}
@@ -338,7 +338,7 @@ namespace gezi {
 
 		void Densify(value_type maxSparsity)
 		{
-			if (length > 0 && (keepDense || Count() >= (uint64)(length * maxSparsity)))
+			if (length > 0 && (keepDense || (uint64)Count() >= (uint64)(length * maxSparsity)))
 			{
 				ToDense();
 			}
